% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPE_geocode.R
\name{GPE_geocode}
\alias{GPE_geocode}
\title{GPE_geocode}
\usage{
GPE_geocode(data, address, key)
}
\arguments{
\item{data}{a dataframe containing an address column}

\item{address}{the name of a column in the geocode dataframe containing addresses}

\item{key}{an API key obtained at https://cloud.google.com/maps-platform/}
}
\description{
A thin wrapper for ggmap::mutate_geocode. Takes a dataframe, adds lon/lat columns based on an existing address column using Google's geocoding API (https://developers.google.com/maps/documentation/geocoding/start).
}
\examples{
\dontrun{
GPE_geocode(data = schools, addres = address, key = mykey)
}
}
