% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPE_get_basemap.R
\name{GPE_get_basemap}
\alias{GPE_get_basemap}
\title{GPE_get_basemap}
\usage{
GPE_get_basemap(participants, locations)
}
\arguments{
\item{participants}{a dataframe a dataframe containing "lon" and "lat" columns with WGS84 (Mercator) coordinates}

\item{locations}{a dataframe a dataframe containing "lon" and "lat" columns with WGS84 (Mercator) coordinates}
}
\description{
Get a basemap from Stamen Maps (http://maps.stamen.com/)
}
\examples{
get_basemap(participants = participants_sf, locations = location_df)
}
