% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPE_plot_travel.R
\name{GPE_plot_travel}
\alias{GPE_plot_travel}
\title{GPE_plot_travel}
\usage{
GPE_plot_travel(visits_timedist, add_data_from = NA,
  plot_attribute = NA, plot_type = "histogram", metric = "time")
}
\arguments{
\item{visits_timedist}{a dataframe of "visits" with travel time and distance (see ?GPE_travel_time_dist)}

\item{add_data_from}{(optional) a participants or locations dataframe, containing either a "participant_id" or a "location_id" column, plus a column with an attribute to plot. Must be paired with "plot_attribute"}

\item{plot_attribute}{(optional) the column name from the "add_data" dataframe containing the attribute to show on the plot}

\item{plot_type}{(optional) either "histogram" (default), "boxplot"}

\item{metric}{(optional) measure for travel, either "time" (default), or "distance"}
}
\description{
Plot a visualization showing participant travel time
}
\examples{
GPE_plot_travel(visits_timedist = visits_timedist)
}
