% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPE_summary.R
\name{GPE_summary}
\alias{GPE_summary}
\title{GPE_summary}
\usage{
GPE_summary(visits)
}
\arguments{
\item{visits}{a dataframe of "visits", events when a participant visited a location, containing "participant_id" and "location_id" columns}
}
\description{
Plot a map showing participants, and program locations
}
\examples{
GPE_summary(visits = visits)
}
