% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPE_travel_time_dist.R
\name{GPE_travel_time_dist}
\alias{GPE_travel_time_dist}
\title{GPE_travel_time_dist}
\usage{
GPE_travel_time_dist(visits, participants, locations, key,
  travel_mode = "transit")
}
\arguments{
\item{visits}{a dataframe of "visits", events when a participant visited a location, containgn "participant_id" and location_id columns}

\item{participants}{a dataframe containing "lon" and "lat" columns with WGS84 (Mercator) coordinates and a "participant_id" column}

\item{locations}{a dataframe containing "lon" and "lat" columns with WGS84 (Mercator) coordinates and a "location_id" column}

\item{key}{an API key obtained at https://cloud.google.com/maps-platform/}

\item{travel_mode}{(optional) either "transit" (default), "driving", "walking", or "bicycling"}
}
\description{
Estimate travel time and distance between origins and destinations
}
\examples{
\donttest{
#'GPE_travel_time_dist(visits, participants, locations, key, travel_mode = "walking")
}
}
