% Generated by roxygen2 (4.0.1): do not edit by hand
\name{iid}
\alias{iid}
\title{IID random effects kernel}
\usage{
iid(columns)
}
\arguments{
\item{columns}{A string vector giving the names of features on which the
kernel acts. These will be used to access columns from a dataframe when the
kernel is later used in a GP model.}
}
\value{
A kernel object for which there are a range of associated functions.
}
\description{
Construct an iid 'random effect' kernel.
}
\details{


The iid kernel takes the form:
\deqn{k_{iid}(\mathbf{x}, \mathbf{x}') = \sigma^2 \mathbf{x} \mathbf{x}'}
where \eqn{\mathbf{x}} are indicator variables with each column containing
1s for records in a given group and 0s otherwise and \eqn{\sigma^2} is the
overall variance. This is equivalent to a hierarchical or random-effects
model, with the hierarchical variance given by \eqn{\sigma^2}.

In practice, the active column should actually be a single factor and the
indicator variables will be built internally.
}
\examples{
# add a discrete variable to the pressure dataset
pressure$group <- factor(sample(letters[1:3], 19, replace = TRUE))

# construct an iid kernel over this
k1 <- iid('group')

# evaluate and visualise it
image(k1(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{composition}}, \code{\link{kron}},
  \code{\link{prod.kernel}}, \code{\link{sum.kernel}};
  \code{\link{lin}}; \code{\link{per}}; \code{\link{rbf}}
}

