% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kernel}
\alias{is.kernel}
\alias{kernel}
\alias{print.kernel}
\alias{summary.kernel}
\title{kernel object}
\usage{
is.kernel(x)

\method{print}{kernel}(x, ...)

\method{summary}{kernel}(object, digits = getOption("digits"))
}
\arguments{
\item{kernel}{A kernel object}

\item{\dots}{For compatibility with the generic print function, not used.}

\item{digits}{The number of digits to display for the kernel parameters}
}
\description{
Generic functions associated with the kernel class.
}
\examples{
# construct a kernel with one feature
k1 <- rbf('temperature')
# is it a kernel?
is.kernel(k1)
# print the kernel's basic structure
print(k1)

# this is also the default action for displaying the object:
k1
# a more detailed summary of the kernel's structure
summary(k1)
}

