% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lin}
\alias{lin}
\title{Linear kernel}
\usage{
lin(columns)
}
\arguments{
\item{columns}{A string vector giving the names of features on which the
kernel acts. These will be used to access columns from a dataframe when the
kernel is later used in a GP model.}
}
\value{
A kernel object for which there are a range of associated functions.
}
\description{
Construct a linear kernel.
}
\details{


The linear kernel takes the form:
\deqn{k_{lin}(\mathbf{x}, \mathbf{x}') = \sigma_b^2 + \sigma_v^2 (\mathbf{x} - c)(\mathbf{x}' - c)}
where \eqn{\mathbf{x}} are the covariates on which the kernel is active,
\eqn{c} determines the value(s) of \eqn{x} through which all realisations pass,
\eqn{\sigma_b^2} is a prior over the absolute value of the intercept and
\eqn{\sigma_v^2} isa  prior over the slopes of the realisations.
}
\examples{
# construct a kernel with one feature
k1 <- lin('temperature')

# and another with two features
k2 <- lin(c('temperature', 'pressure'))

# evaluate them on the pressure dataset
image(k1(pressure))
image(k2(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{composition}}, \code{\link{kron}},
  \code{\link{prod.kernel}}, \code{\link{sum.kernel}};
  \code{\link{rbf}}
}

