% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\docType{data}
\name{locations}
\alias{locations}
\title{Made up locations in Buenos Aires, Argentina}
\format{A data frame with 76 rows and 4 variables:
\describe{
  \item{location_ID}{Unique identifier}
  \item{type}{a categorical variable}
  \item{lon}{Longitude}
  \item{lat}{Latitude}
}}
\usage{
data(participants)
}
\description{
Points sampled at random, useful for testing mapping functions.
}
\keyword{datasets}
