% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kernel_basis_mat32.R
\name{mat32}
\alias{mat32}
\title{Matern 3/2 kernel}
\usage{
mat32(columns, sigma = 1, l = rep(1, length(columns)))
}
\arguments{
\item{columns}{a string vector giving the names of features on which the
kernel acts. These will be used to access columns from a dataframe when the
kernel is later used in a GP model.}

\item{sigma}{a positive scalar parameter giving (the square-root of) the
overall variance of the kernel}

\item{l}{a positive scalar or vector parameter giving the characteristic lengthscale
of the kernel (how rapidly the covariance decays with difference in the
value of the covariate). Larger values of \code{l} imply less complex
functions.}
}
\value{
a kernel object for which there are a range of associated functions.
}
\description{
Construct a Matern 3/2 kernel.
}
\details{


The Matern 3/2 kernel takes the form:
\deqn{k_{mat32}(\mathbf{x}, \mathbf{x}') = \sigma^2 (1 + \sqrt{3} \\ \mathbf{dist}) exp(-\sqrt{3} \\ \mathbf{dist})}
\deqn{\mathbf{dist} = {\sum\limits_{d=1}^D \left(\frac{(x_d - x_d')}{2l_d^2}\right)}^2}
where \eqn{\mathbf{x}} are the covariates on which the kernel is active, \eqn{l_d}
are the characteristic lengthscales for each covariate (column) \eqn{x_d}
and \eqn{\sigma^2} is the overall variance.
Larger values of \eqn{l_i} correspond to functions in which change less
rapidly over the values of the covariates.
}
\examples{
# construct a kernel with one feature
k1 <- mat32('temperature')

# and another with two features
k2 <- mat32(c('temperature', 'pressure'))

# evaluate them on the pressure dataset
image(k1(pressure))
image(k2(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{^.kernel}}, \code{\link{composition}},
  \code{\link{kron}}, \code{\link{prod.kernel}},
  \code{\link{sum.kernel}}; \code{\link{expo}};
  \code{\link{iid}}; \code{\link{int}}; \code{\link{lin}};
  \code{\link{mat52}}; \code{\link{per}}; \code{\link{rbf}}
}

