% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\docType{data}
\name{participants}
\alias{participants}
\title{Made up participants}
\format{A data frame with 7452 rows and 5 variables:
\describe{
  \item{participant_id}{Participant unique ID}
  \item{group}{An example categorical attribute}
  \item{lon}{Participant longitude - home address or other origin point}
  \item{lat}{Participant latitude - home address or other origin point}
  \item{income_bracket}{An example ordinal attribute}
}}
\usage{
data(participants)
}
\description{
Synthetica data describing "participants", i.e. public program beneficiaries or similar agents.
Useful for testing aggregation and analysis functions.
}
\keyword{datasets}
