% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kernel_basis_per.R
\name{per}
\alias{per}
\title{Periodic kernel}
\usage{
per(columns, p = 1, l = 1, sigma = 1)
}
\arguments{
\item{columns}{a string vector giving the names of features on which the
kernel acts. These will be used to access columns from a dataframe when the
kernel is later used in a GP model.}

\item{p}{a positive scalar parameter giving the periodicity of the kernel,
the difference between values of the covariate at which period signal
repeats}

\item{l}{a positive scalar or vector parameter giving the characteristic lengthscale
of the kernel (how rapidly the covariance decays with difference in the
value of the covariate). Larger values of \code{l} imply less complex
functions.}

\item{sigma}{a positive scalar parameter giving (the square-root of) the
overall variance of the kernel}
}
\value{
A kernel object for which there are a range of associated functions, see \code{\link{kernel}} and \code{\link{access}} for details.
}
\description{
Construct a periodic kernel.
}
\details{


The periodic kernel takes the form:
\deqn{k_{per}(\mathbf{x}, \mathbf{x}') = \sigma^2 exp \left(-\frac{2sin^2(\pi | \mathbf{x} - \mathbf{x}' | /p)}{l^2} \right)}
where \eqn{\mathbf{x}} are the covariates on which the kernel is active,
\eqn{p} determines the periodicity (distance between successive peaks),
\eqn{l} is a characteristic lengthscale, as in the rbf kernel, and \eqn{\sigma^2}
is the amplitude of the signal
}
\examples{
# construct a kernel with one feature
k1 <- per('temperature')

# and another with two features
k2 <- per(c('temperature', 'pressure'))

# evaluate them on the pressure dataset
image(k1(pressure))
image(k2(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{^.kernel}}, \code{\link{composition}},
  \code{\link{kron}}, \code{\link{prod.kernel}},
  \code{\link{sum.kernel}}; \code{\link{expo}};
  \code{\link{iid}}; \code{\link{int}}; \code{\link{lin}};
  \code{\link{mat32}}; \code{\link{mat52}};
  \code{\link{rbf}}; \code{\link{rq}}
}

