% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rbf}
\alias{rbf}
\title{Radial basis function kernel}
\usage{
rbf(columns)
}
\arguments{
\item{columns}{A string vector giving the names of features on which the
kernel acts. These will be used to access columns from a dataframe when the
kernel is later used in a GP model.}
}
\value{
A kernel object for which there are a range of associated functions.
}
\description{
Construct a radial basis function (A.K.A. squared-exponential) kernel.
}
\details{


The rbf kernel takes the form:
\deqn{k_{rbf} = \sigma^2 exp(-\frac{{(x - x')}^2}{2l^2})}
}
\examples{
# construct a kernel with one feature
k1 <- rbf('temperature')

# and another with two features
k2 <- rbf(c('temperature', 'pressure'))

# evaluate them on the pressure dataset
image(k1(pressure))
image(k2(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{composition}}, \code{\link{kron}},
  \code{\link{prod.kernel}}, \code{\link{sum.kernel}}
}

