% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_basis_rq.R
\name{rq}
\alias{rq}
\title{Rational quadratic kernel}
\usage{
rq(columns, sigma = 1, l = rep(1, length(columns)), alpha = 1)
}
\arguments{
\item{columns}{a string vector giving the names of features on which the
kernel acts. These will be used to access columns from a dataframe when the
kernel is later used in a GP model.}

\item{sigma}{a positive scalar parameter giving (the square-root of) the 
overall variance of the kernel}

\item{l}{a positive scalar or vector parameter giving the characteristic lengthscale 
of the kernel (how rapidly the covariance decays with difference in the 
value of the covariate). Larger values of \code{l} imply less complex 
functions.}

\item{alpha}{a positive scalar or vector parameter controlling the roughness
of the GPs drawn from the kernel. Larger values of \code{alpha} imply 
rougher complex functions.}
}
\value{
A kernel object for which there are a range of associated functions, see \code{\link{kernel}} and \code{\link{access}} for details.
}
\description{
Construct a rational quadratic kernel.
}
\details{
The rational quadratic kernel takes the form:
\deqn{k_{rq}(\mathbf{x}, \mathbf{x}') = \sigma^2 \left(1 + \frac{\mathbf{r} ^ 2}{\alpha} \right) ^ {-\alpha}}
\deqn{\mathbf{r} = {\sqrt{\sum\limits_{d=1}^D \left(\frac{(x_d - x_d')}{2l_d^2}\right) ^ 2}}}
where \eqn{\mathbf{x}} are the covariates on which the kernel is active, 
\eqn{l_d} are the characteristic lengthscales for each covariate (column) 
\eqn{x_d}, \eqn{\sigma^2} is the overall variance and \eqn{\alpha} controls the
roughness.

Larger values of \eqn{l_i} correspond to functions in which change less 
rapidly over the values of the covariates.
}
\examples{
# construct a kernel with one feature
k1 <- rq('temperature')

# and another with two features
k2 <- rq(c('temperature', 'pressure'))

# evaluate them on the pressure dataset
image(k1(pressure))
image(k2(pressure))

}
\seealso{
Other kernel.constructors: \code{\link{composition}},
  \code{\link{expo}}, \code{\link{iid}}, \code{\link{int}},
  \code{\link{lin}}, \code{\link{mat32}},
  \code{\link{mat52}}, \code{\link{per}}, \code{\link{rbf}}
}

