% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\docType{data}
\name{schools}
\alias{schools}
\title{Ten public schools in Buenos Aires, Argentina}
\format{A data frame with 10 rows and r variables:
\describe{
  \item{name}{School name}
  \item{address}{School address}
  \item{level}{Education level\(s\) offered by the school}
}}
\source{
\url{https://data.buenosaires.gob.ar/}
}
\usage{
schools
}
\description{
A small dataset basic information of ten public schools in Buenos Aires, Argentina.
}
\keyword{datasets}
