% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits.R
\docType{data}
\name{visits}
\alias{visits}
\title{Made up visits}
\format{A data frame with 19123 rows and 3 variables:
\describe{
  \item{date}{Visit date}
  \item{participant_id}{Participant unique ID}
  \item{location_id}{Location Unique ID}
}}
\usage{
visits
}
\description{
Randomly selected pairs of "participants" and "locations", representing visits,
useful for testing visit aggregation and analysis functions.
}
\keyword{datasets}
