% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits_timedist.R
\docType{data}
\name{visits_timedist}
\alias{visits_timedist}
\title{Made up visits, with trip duration and distance}
\format{A data frame with 234 rows and 6 variables:
\describe{
  \item{date}{Visit date}
  \item{participant_id}{Participant unique ID}
  \item{location_id}{Location Unique ID}
  \item{time_minutes}{trip duration in minutes}
  \item{distance_km}{trip distance in km}
  \item{mode}{transportation mode -transit, walking, driving, or bycicle}
}}
\usage{
visits_timedist
}
\description{
Randomly selected pairs of "participants" and "locations", representing visits,
useful for testing visit aggregation and analysis functions.
}
\keyword{datasets}
