% Generated by roxygen2 (4.0.1): do not edit by hand
\name{access}
\alias{access}
\alias{getColumns}
\alias{getParameters}
\alias{getSubKernel}
\alias{getType}
\alias{setParameters}
\title{Interaction with kernel objects}
\usage{
getType(kernel)

getColumns(kernel)

getParameters(kernel)

getSubKernel(kernel, which = 1)

setParameters(kernel, parameter_list)
}
\arguments{
\item{kernel}{A kernel object}

\item{which}{Which of the two \emph{immediate} subkernels of the kernel to extract}

\item{parameter_list}{A named list with names giving the parameters to be
updated and values giving the new values they should take on.}
}
\description{
Access and interact with kernel objects.
These functions allow you to report the type of kernel (\code{getType}),
view the overall kernel structure (\code{?}), extract subkernels (\code{getSubKernel})
report kernel parameters (\code{getParameters}) and update them (\code{setParameters}).
}
\examples{
# construct a kernel with one feature
k1 <- rbf('temperature')
# get the kernel type
getType(k1)
# get the names of the columns the kernel acts on
getColumns(k1)
# get the parameters of the kernel
params <- getParameters(k1)
params
# build a compositional kernel
k2 <- k1 + k1 * k1

# extract a subkernel
k3 <- getSubKernel(k2, 1)
# evaluate and visualise it
image(k1(pressure))

# change the length scale
params$l <- 100
k2 <- setParameters(k1, params)

# evaluate and visualise the new kernel
image(k1(pressure))
}

