% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gp}
\alias{gp}
\title{Gaussian process models}
\usage{
gp(response, kernel, data, family = gaussian, mean_function = NULL,
  inducing_data = NULL, inference = c("default", "full", "FITC", "Laplace"),
  verbose = FALSE)
}
\arguments{
\item{response}{A vector or matrix containing values of the response
variable to be modelled.}

\item{kernel}{A kernel object.}

\item{data}{A data frame containing the covariates against which to model
the response variable. This must have the same number of rows as
\code{response} and contain named variables matching those referred to by
\code{kernel}.}

\item{family}{A \code{\link{family}} object giving the likelihood and link
function to be used to fit the model. Currently only \code{gaussian},
\code{poisson} and \code{binomial} (only Bernoulli) are supported.}

\item{mean_function}{An optional function specifying the prior over the mean
of the gp, in other words a 'first guess' at what the true function is.
This must act on a dataframe with named variables matching some of those in
\code{data} and return a vector giving a single value for each row in the
dataframe. Note that this function must return a prediction on the scale of
the link, rather than the response. If \code{NULL} then a prior mean is
assumed to be 0 for all observations.}

\item{inducing_data}{An optional dataframe containing the locations of
inducing points to be used when carrying out sparse inference (e.g. FITC).
This must contain variables with names matching those referenced by
\code{kernel} and \code{mean_function}. This should have fewer rows than
\code{data} and \code{response}.}

\item{inference}{A string specifying the inference method to be used to
estimate the values of the latent parameters. If \code{'default'} an
appropriate method is picked for the likelihood specified. See details
section for the list of default inference methods.}
}
\value{
A fitted gp object for which there aren't yet any associated
functions. But there will be.
}
\description{
Fiting (latent) Gaussian process models using a range of
inference methods.
}
\details{
The default inference method for a model with the family
\code{gaussian(link = 'identity')} is full, direct inference, for
\code{binomial(link = 'logit')} and \code{binomial(link = 'probit')}
the default is full, Laplace inference (though note that only bernoulli)
data is handled at the moment.
}
\examples{
# make some fake data
n <- 100  # observations
m <- 10  # inducing points

# dataframes
df <- data.frame(x = sort(runif(n, -5, 5)))
inducing_df <- data.frame(x = sort(runif(m, -5, 5)))
prediction_df <- data.frame(x = seq(min(df$x), max(df$x), len = 500))

# fake Gaussian response data
f <- sin(df$x)
y <- rnorm(n, f, 1)

# construct a kernel (including the observation error)
kernel <- rbf('x') + iid()

# fit a full (non-sparse) GP model (without updating the hyperparameters)
# as this is the default
m1 <- gp(y, kernel, df, gaussian)

# fit another with FITC sparsity
m2 <- gp(y, kernel, df, gaussian, inference = 'FITC', inducing_data = df)#inducing_df)
# summary stats, other associated functions still to come

# construct a poisson response variable
y2 <- rpois(n, exp(f))

# fit a GP model by Laplace approximation
# (note no observation error in this model)
m2 <- gp(y2, rbf('x'), df, poisson)
}

