% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPE_plot_map.R
\name{GPE_plot_map}
\alias{GPE_plot_map}
\title{GPE_plot_map}
\usage{
GPE_plot_map(participants, locations = NULL,
  participant_attribute = NA, location_attribute = NA)
}
\arguments{
\item{participants}{a dataframe containing "lon" and "lat" columns with WGS84 (Mercator) coordinates}

\item{locations}{a dataframe containing "lon" and "lat" columns with WGS84 (Mercator) coordinates}

\item{participant_attribute}{(optional) a participants dataframe column containing an attribute to show on the map}

\item{location_attribute}{(optional) a locations dataframe column containing an attribute to show on the map}
}
\description{
Plot a map showing participants, and program locations
}
\examples{
GPE_plot_map(participants = participants, locations = locations)
}
