% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kernel_basis_iid.R
\name{iid}
\alias{iid}
\title{IID random effects kernel}
\usage{
iid(column = NULL, sigma = 1)
}
\arguments{
\item{column}{optionally, a single string giving the name of a \emph{factor}
feature on which the kernel acts. This will be used to access columns from
a dataframe when the kernel is evaluated . If \code{NULL} then the iid
effect acts on each datapoint, rather than on groups.}

\item{sigma}{a positive scalar parameter giving (the square-root of) the
overall variance of the kernel}
}
\value{
a kernel object for which there are a range of associated functions.
}
\description{
Construct an iid normal `random effects' kernel.
}
\details{


The iid kernel takes the form:
\deqn{k_{iid}(\mathbf{x}, \mathbf{x}') = \sigma^2 \mathbf{x} \mathbf{x}'}
where \eqn{\mathbf{x}} are indicator variables with each column containing
1s for records in a given group and 0s otherwise and \eqn{\sigma^2} is the
overall variance. This is equivalent to a hierarchical or random-effects
model: \eqn{z_j ~ N(0, \sigma^2)} with \eqn{j} indexing the groups of the
specified factor.

In practice, the active column should actually be a single factor and the
indicator variables will be built internally.
}
\examples{
# add a discrete variable to the pressure dataset
pressure$group <- factor(sample(letters[1:3], 19, replace = TRUE))

# construct an iid kernel over this
k1 <- iid('group')

# evaluate and visualise it
image(k1(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{composition}}, \code{\link{kron}},
  \code{\link{prod.kernel}}, \code{\link{sum.kernel}};
  \code{\link{expo}}; \code{\link{int}}; \code{\link{lin}};
  \code{\link{mat32}}; \code{\link{mat52}};
  \code{\link{per}}; \code{\link{rbf}}
}

