% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kernel_basis_int.R
\name{int}
\alias{int}
\title{Intercept kernel}
\usage{
int(sigma = 1)
}
\arguments{
\item{sigma}{a positive scalar parameter giving (the square-root of) the
overall variance of the kernel}
}
\value{
a kernel object for which there are a range of associated functions.
}
\description{
Construct an intercept kernel
}
\details{


The intercept kernel takes the form:
\deqn{k_{int}(\mathbf{x}, \mathbf{x}') = \sigma^2}
where \eqn{\sigma^2} is the variance of the prior over the value of the
intercept. This is equivalent to a normal prior over the value of an
intercept term: \eqn{\alpha ~ N(0, \sigma^2)}.
}
\examples{
# construct an intercept kernel
k1 <- int()

# evaluate and visualise it
image(k1(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{^.kernel}}, \code{\link{composition}},
  \code{\link{kron}}, \code{\link{prod.kernel}},
  \code{\link{sum.kernel}}; \code{\link{expo}};
  \code{\link{iid}}; \code{\link{lin}};
  \code{\link{mat32}}; \code{\link{mat52}};
  \code{\link{per}}; \code{\link{rbf}}
}

