% Generated by roxygen2 (4.0.1): do not edit by hand
\name{int}
\alias{int}
\title{Intercept kernel}
\usage{
int()
}
\value{
a kernel object for which there are a range of associated functions.
}
\description{
Construct an intercept kernel
}
\details{


The intercept kernel takes the form:
\deqn{k_{int}(\mathbf{x}, \mathbf{x}') = \sigma^2}
where \eqn{\sigma^2} is the variance of the prior over the value of the
intercept. This is equivalent to a normal prior over the value of an
intercept term: \eqn{\alpha ~ N(0, \sigma^2)}.
}
\examples{
# construct an intercept kernel
k1 <- int()

# evaluate and visualise it
image(k1(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{composition}}, \code{\link{kron}},
  \code{\link{prod.kernel}}, \code{\link{sum.kernel}};
  \code{\link{iid}}; \code{\link{lin}}; \code{\link{per}};
  \code{\link{rbf}}
}

