% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kernel}
\alias{is.kernel}
\alias{kernel}
\alias{print.kernel}
\alias{summary.kernel}
\title{gpe kernel class}
\usage{
is.kernel(x)

\method{print}{kernel}(x, ...)

\method{summary}{kernel}(object, ..., digits = max(3, getOption("digits") -
  3))
}
\arguments{
\item{x}{a kernel object, or some other object to be tested}

\item{\dots}{additional arguments for compatibility with generic functions}

\item{object}{a kernel object}

\item{digits}{the number of digits to display for the kernel parameters}
}
\description{
Generic functions associated with the kernel class.
}
\examples{
# construct a kernel with one feature
k1 <- rbf('temperature')
# is it a kernel?
is.kernel(k1)
# print the kernel's basic structure
print(k1)

# this is also the default action for displaying the object:
k1
# a more detailed summary of the kernel's structure
summary(k1)
}

