% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kernel_basis_lin.R
\name{lin}
\alias{lin}
\title{Linear kernel}
\usage{
lin(columns, sigma = 1, c = rep(0, length(columns)))
}
\arguments{
\item{columns}{a string vector giving the names of features on which the
kernel acts. These will be used to access columns from a dataframe when the
kernel is later used in a GP model.}

\item{sigma}{a positive scalar parameter giving (the square-root of) the
overall variance of the kernel}

\item{c}{a scalar or vector parameter with continuous support giving
the location on the x axis where y = 0 (an x-axis intercept)}
}
\value{
a kernel object for which there are a range of associated functions.
}
\description{
Construct a linear kernel.
}
\details{


The linear kernel takes the form:
\deqn{k_{lin}(\mathbf{x}, \mathbf{x}') = \sigma^2 (\mathbf{x} - c)(\mathbf{x}' - c)}
where \eqn{\mathbf{x}} are the covariates on which the kernel is active,
\eqn{c} determines the value(s) of \eqn{x} through which all realisations
pass and \eqn{\sigma^2} is a prior over the slopes of the realisations.
}
\examples{
# construct a kernel with one feature
k1 <- lin('temperature')

# and another with two features
k2 <- lin(c('temperature', 'pressure'))

# evaluate them on the pressure dataset
image(k1(pressure))
image(k2(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{composition}}, \code{\link{kron}},
  \code{\link{prod.kernel}}, \code{\link{sum.kernel}};
  \code{\link{iid}}; \code{\link{int}}; \code{\link{per}};
  \code{\link{rbf}}
}

