% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kernel_basis_per.R
\name{per}
\alias{per}
\title{Periodic kernel}
\usage{
per(columns)
}
\arguments{
\item{columns}{a string vector giving the names of features on which the
kernel acts. These will be used to access columns from a dataframe when the
kernel is later used in a GP model.}
}
\value{
a kernel object for which there are a range of associated functions.
}
\description{
Construct a periodic kernel.
}
\details{


The periodic kernel takes the form:
\deqn{k_{per}(\mathbf{x}, \mathbf{x}') = \sigma^2 exp \left(-\frac{2sin^2(\pi | \mathbf{x} - \mathbf{x}' | /p)}{l^2} \right)}
where \eqn{\mathbf{x}} are the covariates on which the kernel is active,
\eqn{p} determines the periodicity (distance between successive peaks),
\eqn{l} is a characteristic lengthscale, as in the rbf kernel, and \eqn{\sigma^2}
is the amplitude of the signal
}
\examples{
# construct a kernel with one feature
k1 <- per('temperature')

# and another with two features
k2 <- per(c('temperature', 'pressure'))

# evaluate them on the pressure dataset
image(k1(pressure))
image(k2(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{composition}}, \code{\link{kron}},
  \code{\link{prod.kernel}}, \code{\link{sum.kernel}};
  \code{\link{iid}}; \code{\link{int}}; \code{\link{lin}};
  \code{\link{rbf}}
}

