% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rbf}
\alias{rbf}
\title{Radial basis function kernel}
\usage{
rbf(columns)
}
\arguments{
\item{columns}{A string vector giving the names of features on which the
kernel acts. These will be used to access columns from a dataframe when the
kernel is later used in a GP model.}
}
\value{
A kernel object for which there are a range of associated functions.
}
\description{
Construct a radial basis function (A.K.A. squared-exponential) kernel.
}
\details{


The rbf kernel takes the form:
\deqn{k_{rbf}(\mathbf{x}, \mathbf{x}') = \sigma^2 exp\left[-\frac{1}{2} {\sum\limits_{d=1}^D \left(\frac{(x_d - x_d')}{2l_d^2}\right)}^2\right]}
where \eqn{\mathbf{x}} are the covariates on which the kernel is active, \eqn{l_d}
are the characteristic lengthscales for each covariate (column) \eqn{x_d}
and \eqn{\sigma^2} is the overall variance.

Larger values of \eqn{l_i} correspond to functions in which change less
rapidly over the values of the covariates.
}
\examples{
# construct a kernel with one feature
k1 <- rbf('temperature')

# and another with two features
k2 <- rbf(c('temperature', 'pressure'))

# evaluate them on the pressure dataset
image(k1(pressure))
image(k2(pressure))
}
\seealso{
Other kernel.constructors: \code{\link{*.kernel}},
  \code{\link{+.kernel}}, \code{\link{\%x\%}},
  \code{\link{composition}}, \code{\link{kron}},
  \code{\link{prod.kernel}}, \code{\link{sum.kernel}};
  \code{\link{iid}}; \code{\link{int}}; \code{\link{lin}};
  \code{\link{per}}
}

